Função de edição de produtos pelo admin adicionada

+++ react-app/src/components/deleteProduct.jsx
@@ -1,24 +1,23 @@
 import React, { useState, useEffect } from 'react';
 import Searchbar from './Searchbar';
 import SearchResultsList from './searchResultsList';
-import { useNavigate } from 'react-router-dom';
 
 const DeleteProduct = () => {
   const [results, setResults] = useState([]);
-  const [products, setProducts] = useState([]);
-  const navigate = useNavigate();
 
-  const handleClick = (productId) => {
-    fetch(`http://localhost:7000/products/${productId}`, {
-      method: 'DELETE',
-    })
-      .then(() => {
-        // Refresh the results after successful deletion
-        fetchResults();
+  const handleClick = (action, productId) => {
+    if (action === 'delete') {
+      fetch(`http://localhost:7000/products/${productId}`, {
+        method: 'DELETE',
       })
-      .catch(error => console.log(error));
+        .then(() => {
+          console.log('Product deleted');
+          setResults(prevResults => prevResults.filter(result => result.id !== productId));
+        })
+        .catch(error => console.log(error));
+    } 
   };
-
+  
   const fetchResults = () => {
     fetch('http://localhost:7000/products')
       .then(response => response.json())
